NavigAIte/backend/requirements.txt
fastapi
uvicorn
openai
gspread
google-auth
oauth2client
pdfplumber
python-dotenv
requests
pypdf2 
tiktoken
